also probe target for load-test before sending bulk requests
